Formatting the size into human readable strings.
